图,json,分屏数据同步完成 #26
