+++ ui/src/pages/workflow/app-design/app-design.vue
@@ -8,10 +8,13 @@ import {
   isEndpointNode,
   convertNodeType,
   generateComponentList,
-  generateDefaultPropertiesStatus,
+  generateFormData,
+  findComponentByType,
   generateFormFields,
   generateStaticFormFields,
   generateStaticFormData,
+  mapFlowDataModelToRuleGoModel,
+  mapRuleGoModelToFlowDataModel,
 } from '@src/pages/workflow/app-design/utils';
 import {
   NODE_DEFAULT_HEIGHT,
@@ -25,12 +28,25 @@ import RunDrawer from '@src/pages/workflow/app-design/run-drawer/run-drawer.vue'
 import NodeList from '@src/pages/workflow/app-design/node-list/node-list.vue';
 import NodeForm from '@src/pages/workflow/app-design/node-form/node-form.vue';
 
+/**
+ * remark:ReversalMinute
+ * 以数据驱动的方式进行控制
+ */
 const props = defineProps({
-  flowData: {
+  /**
+   * ruleGo model
+   */
+  modelValue: {
     type: Object,
     default: () => ({}),
   },
 });
+/**
+ * ruleGo model
+ */
+const emit = defineEmits(['update:modelValue']);
+
+
 
 const closeNodeFormBus = EventBus.closeNodeForm();
 const showNodeMenuBus = EventBus.showNodeMenu();
@@ -84,6 +100,7 @@ const fields = ref({
 const nextNodes = ref([]);
 const selectedNodeModel = ref(undefined);
 const anchorCanConnectOnlyOneNode = ref(false);
+
 const { pause, resume } = watchPausable(
   () => formData.value,
   () => {
@@ -126,378 +143,40 @@ function generateStartMenuItem() {
 }
 
 function clearFlowData() {
+  if (!flowViewRef.value) return;
+
+  // 清空数据
   nodes.value = [];
   edges.value = [];
-  flowViewRef.value?.getLf().clearData();
+  flowViewRef.value.getLf().clearData();
+
+}
+
+function rerenderFlowData() {
+  if (!flowViewRef.value) return;
+  const lf = flowViewRef.value.getLf();
+  // 保存当前视图的偏移和缩放状态
+  const userData = cloneDeep(props.modelValue);
+  const { nodes, edges } = mapRuleGoModelToFlowDataModel(userData, menuList.value);
+
+  //  重新渲染数据
+  lf.renderRawData({
+    nodes,
+    edges
+  })
 }
 
 function generateFlowData() {
   if (!flowViewRef.value) return;
 
   // 接口请求的数据
-  const userData = cloneDeep(props.flowData);
-  let tNodes = [];
-  let tEdges = [];
-
-  // 转换逻辑
-
-  let ruleChainAdditionalInfo = {};
-  //规则链Id名称等数据
-  if (userData.ruleChain) {
-    ruleChainAdditionalInfo = userData.ruleChain.additionalInfo || {};
-  }
-  //默认数据
-  let initData = {
-    //开始节点ID
-    startNodeId: 'start',
-    startNodeText: '开始',
-    startNodeType: 'start',
-    nodeCount: 0,
-    edgeCount: 0,
-    nodeWidth: 200,
-  };
-  //开始节点
-  const options = {
-    startX: 280,
-    startY: 280,
-  };
-  let referenceX =
-    parseInt(ruleChainAdditionalInfo.layoutX || options.startX) ||
-    options.startX;
-  let referenceY =
-    parseInt(ruleChainAdditionalInfo.layoutY || options.startY) ||
-    options.startY;
-  let hasEndpoint = userData?.metadata?.endpoints?.length > 0;
-  //如果没有endpoint节点，添加start固定节点
-  if (!hasEndpoint) {
-    tNodes.push({
-      id: initData.startNodeId,
-      type: initData.startNodeType,
-      x: referenceX,
-      y: referenceY,
-      text: initData.startNodeText,
-    });
-  }
-
-  //第一个普通节点(非endpoint节点)
-  let firstNode = null;
-  //最大节点Id序号
-  let maxNodeIdSeq = 0;
-  let firstNodeIndex =
-    (userData.metadata && userData.metadata.firstNodeIndex) || 0;
-  //转换节点
-  if (userData && userData.metadata && userData.metadata.nodes) {
-    userData.metadata.nodes.forEach((item, index) => {
-      initData.nodeCount++;
-      //获取节点类型
-      let nodeType = item.type;
-
-      if (firstNodeIndex === index) {
-        firstNode = item;
-      }
-      referenceX = referenceX + 150;
-      let additionalInfo = item.additionalInfo || {};
-      let x = additionalInfo.layoutX || referenceX;
-      let y = additionalInfo.layoutY || referenceY;
-
-      tNodes.push({
-        id: item.id,
-        type: nodeType,
-        x: x,
-        y: y,
-        text: item.name,
-        properties: {
-          // view: nodeComponents[item.type],
-          model: item,
-        },
-      });
-    });
-  }
-  //转换输入端节点
-  if (hasEndpoint) {
-    userData.metadata.endpoints.forEach((item, index) => {
-      referenceX = referenceX + 150;
-      let additionalInfo = item.additionalInfo || {};
-      let x = additionalInfo.layoutX || referenceX;
-      let y = additionalInfo.layoutY || referenceY;
-      tNodes.push({
-        id: item.id,
-        type: item.type,
-        x: x,
-        y: y,
-        text: item.name,
-        properties: {
-          // view: nodeComponents[item.type],
-          model: item,
-        },
-      });
-    });
-  }
-  //输入端节点连接列表
-  let endpointConnections = [];
-  //如果没有endpoint节点，添加start节点和第一个节点连线
-  if (!hasEndpoint && firstNode) {
-    endpointConnections = [
-      { fromId: initData.startNodeId, toId: firstNode.id },
-    ];
-  }
-
-  //处理输入端节点连线
-  if (userData.metadata && userData.metadata.endpoints) {
-    userData.metadata.endpoints.forEach((item, index) => {
-      endpointConnections = endpointConnections.concat(
-        getEndpointConnections(item.id, item, firstNode),
-      );
-    });
-  }
-
-  let connections = endpointConnections;
-  //转换其他节点边连线
-  if (userData && userData.metadata && userData.metadata.connections) {
-    connections = connections.concat(userData.metadata.connections);
-  }
-  //渲染边
-  connections.forEach((item) => {
-    initData.edgeCount++;
-    let edge = createEdge(initData, tNodes, item);
-    tEdges.push(edge);
-  });
-
-  tNodes = tNodes.map((item) => {
-    return generateFlowNode(item, connections);
-  });
-
-  tEdges = tEdges.map((item) => {
-    return generateFlowEdge(item);
-  });
+  const userData = cloneDeep(props.modelValue);
+  const { nodes: tNodes, edges: tEdges } = mapRuleGoModelToFlowDataModel(userData, menuList.value);
 
   nodes.value = tNodes;
   edges.value = tEdges;
 
   flowViewRef.value.lfRender();
-}
-//创建边
-function createEdge(initData, nodes, item) {
-  let edge = {
-    id: 'edge_' + initData.edgeCount,
-    type: 'custom-bezier',
-    sourceNodeId: item.fromId,
-    targetNodeId: item.toId,
-    startPoint: {},
-    endPoint: {},
-    text: item.label || '',
-    properties: {
-      model: { ...item },
-    },
-  };
-  let sourceNode = getNodeByID(nodes, edge.sourceNodeId);
-  let targetNode = getNodeByID(nodes, edge.targetNodeId);
-  if (sourceNode) {
-    edge.startPoint.x = sourceNode.x + initData.nodeWidth / 2;
-    edge.startPoint.y = sourceNode.y;
-  }
-  if (targetNode) {
-    edge.endPoint.x = targetNode.x - initData.nodeWidth / 2 - 10;
-    edge.endPoint.y = targetNode.y;
-  }
-  return edge;
-}
-//通过id查找节点
-function getNodeByID(nodes, id) {
-  let node = {};
-  if (nodes) {
-    nodes.forEach((item) => {
-      if (item.id === id) {
-        node = item;
-      }
-    });
-  }
-  return node;
-}
-//获取endpoint连线
-function getEndpointConnections(fromId, endpointModel, firstNode) {
-  let connections = [];
-  if (!firstNode) {
-    return [];
-  }
-  if (endpointModel.routers && endpointModel.routers.length > 0) {
-    endpointModel.routers.forEach((item) => {
-      if (item.to && item.to.path) {
-        //格式:chainId:nodeId1:nodeId2:nodeId3
-        let values = item.to.path.split(':');
-        let path = toFromPath(item).trim();
-        if (values.length <= 1) {
-          connections.push({
-            fromId: fromId,
-            toId: firstNode.id,
-            routerId: item.id,
-            nodeType: 'endpoint-node',
-            type: path,
-          });
-        } else {
-          //从1开始遍历节点ID
-          for (let i = 1; i < values.length; i++) {
-            connections.push({
-              fromId: fromId,
-              toId: values[i].trim(),
-              nodeType: 'endpoint-node',
-              routerId: item.id,
-              type: path,
-            });
-          }
-        }
-      }
-    });
-  }
-  return connections;
-}
-
-function toFromPath(item) {
-  if (!item.from) {
-    return '';
-  }
-  let params = item.params ? item.params.join(' ') : '';
-  return params + ' ' + (item.from && item.from.path);
-}
-
-function findComponentByType(type) {
-  if (isEmpty(menuList.value)) {
-    return undefined;
-  }
-  const allComponents = Object.keys(menuList.value)
-    .map((key) => {
-      return menuList.value[key]?.components || [];
-    })
-    .flat();
-  const component = allComponents.find((item) => item.type === type);
-  return component;
-}
-
-function generateFlowNode(config, connections) {
-  const configType = config.type;
-  const configAdditionalInfo = config?.properties?.model?.additionalInfo || {};
-  const configConfiguration = config?.properties?.model?.configuration || {};
-  let routers = config?.properties?.model?.routers || [];
-  const configName = config.text;
-  let flowNodeType = convertNodeType(configType);
-  let flowNodeX = config.x || 0;
-  let flowNodeY = config.y || 0;
-  if (configType !== 'start') {
-    flowNodeX = configAdditionalInfo.layoutX
-      ? Number(configAdditionalInfo.layoutX)
-      : 0;
-    flowNodeY = configAdditionalInfo.layoutY
-      ? Number(configAdditionalInfo.layoutY)
-      : 0;
-  }
-
-  if (!flowNodeType) return undefined;
-
-  const defaultNode = findComponentByType(configType);
-  let defaultNodeForm = generateFormData(defaultNode);
-  // defaultNodeForm = merge(defaultNodeForm, configConfiguration);
-  defaultNodeForm = configConfiguration;
-  // defaultNodeForm.title = configName;
-  defaultNodeForm.additionalInfo = configAdditionalInfo || {};
-  defaultNodeForm.additionalInfo.title =
-    config?.properties?.model?.name || configName;
-
-  let defaultNodeFields = generateFormFields(defaultNode?.fields);
-  defaultNodeFields = generateStaticFormFields(
-    defaultNodeFields,
-    config,
-    defaultNode,
-  );
-  let defaultAnchors = generateStaticAnchors(defaultNode);
-  routers = routers.map((item) => {
-    return {
-      id: item.id,
-      path: item.from.path,
-      fromProcessors: item.from.processors,
-      toProcessors: item.to.processors,
-    };
-  });
-
-  const flowNode = {
-    id: config.id,
-    type: flowNodeType,
-    properties: {
-      formData: { ...defaultNodeForm },
-      fields: defaultNodeFields,
-      nextNodes: [],
-      anchors: defaultAnchors,
-      status: generateDefaultPropertiesStatus(),
-      width: configAdditionalInfo.width || NODE_DEFAULT_WIDTH,
-      height: configAdditionalInfo.height || NODE_DEFAULT_HEIGHT,
-      rawNodeType: configType,
-    },
-    width: configAdditionalInfo.width || NODE_DEFAULT_WIDTH,
-    height: configAdditionalInfo.height || NODE_DEFAULT_HEIGHT,
-    x: flowNodeX,
-    y: flowNodeY,
-  };
-  if (flowNodeType) {
-    flowNode.properties.formData = {
-      ...flowNode.properties.formData,
-      routers,
-    };
-  } else {
-    delete flowNode.properties.formData.routers;
-  }
-
-  if (flowNodeType === 'msg-type-switch') {
-    const msgTypeSwitchConnections = (connections || []).filter((item) => {
-      return item.fromId === flowNode.id;
-    });
-    const msgTypeSwitchAnchors = msgTypeSwitchConnections.map((item) => {
-      return {
-        id: item.type,
-        type: 'output',
-        name: item.type,
-        isStatic: false,
-        top: 0,
-      };
-    });
-    let routers = msgTypeSwitchConnections.map((item) => item.type);
-    routers = unionBy(routers);
-    routers = configAdditionalInfo?.routers || routers;
-    flowNode.properties.anchors = msgTypeSwitchAnchors;
-    flowNode.properties.formData.routers = routers;
-  }
-
-  if (flowNodeType === 'switch') {
-    flowNode.properties.anchors.push({
-      x: 0,
-      y: 0,
-      id: 'Default',
-      type: 'output',
-      name: 'Default',
-      isStatic: true,
-    });
-    flowNode.properties.formData.cases = generateSwitchNodeFormValue(
-      flowNode?.properties?.formData?.cases || [],
-    );
-  }
-
-  return flowNode;
-}
-
-function generateFlowEdge(config) {
-  const id = config.id || `edge:${nanoid()}`;
-  const type = config?.properties?.model?.type;
-  const sourceNodeId = config.sourceNodeId;
-  const sourceAnchorId = type;
-  const targetNodeId = config.targetNodeId;
-  const targetAnchorId = `${targetNodeId}_input`;
-  return {
-    id,
-    type: 'custom-bezier',
-    sourceNodeId,
-    sourceAnchorId,
-    targetNodeId,
-    targetAnchorId,
-    text: config.text,
-  };
 }
 
 function anchorClickHandler({ clientX, clientY }) {
@@ -526,6 +205,7 @@ function flowNodeUpdateNodePropertiesHandler(data) {
   formData.value = data.formData;
   fields.value = data.fields;
   nextNodes.value = data.nextNodes;
+  handleEmitUpdate();
   setTimeout(() => {
     resume();
   }, 0);
@@ -563,46 +243,12 @@ function generateSwitchNodeFormValue(val) {
   });
 }
 
-function generateFormData(config) {
-  if (!config) return {};
-  let form = {
-    // title: config.label,
-    additionalInfo: {
-      title: config.label,
-      description: '',
-    },
-  };
-  (config.fields || []).forEach((field) => {
-    let value = cloneDeep(field.defaultValue);
-    form[field.name] = value;
-  });
-
-  return form;
-}
-
-function generateStaticAnchors(config) {
-  let anchors = [];
-  if (!config) return anchors;
-  const relationTypes = config.relationTypes || [];
-  relationTypes.forEach((relationType) => {
-    return anchors.push({
-      type: 'output',
-      id: relationType.value,
-      name: relationType.label,
-      top: 0,
-      isStatic: true,
-    });
-  });
-
-  return anchors;
-}
-
 function nodeListSelectedHandler(item) {
   const nodeType = item.type;
   let form = generateFormData(item);
   form = generateStaticFormData(form, item);
   let fields = generateFormFields(item.fields);
-  const nodeView = findComponentByType(nodeType);
+  const nodeView = findComponentByType(nodeType, menuList.value);
   fields = generateStaticFormFields(fields, item, nodeView);
   let anchors = generateStaticAnchors(item);
   if (nodeFormRef.value) {
@@ -640,6 +286,7 @@ function nodeListSelectedHandler(item) {
 
   if (addType.value === 'change-node') {
     changeFlowNodeBus.emit(node);
+
   }
 }
 
@@ -697,6 +344,10 @@ function showNodeMenuHandler(clientX, clientY) {
   showNodeList();
 }
 
+/**
+ * 获取逻辑流模型
+ * @returns {{nodes: Array, edges: Array}} 返回逻辑流模型对象，其中包含节点数组和边数组
+ */
 function getData() {
   if (!flowViewRef.value) {
     return {
@@ -709,6 +360,18 @@ function getData() {
 
 closeNodeFormBus.on(nodeFormCloseHandler);
 showNodeMenuBus.on(showNodeMenuHandler);
+
+/**
+* 派发更新事件
+*/
+function handleEmitUpdate() {
+  const logicFlowData = getData();
+  const ruleGoData = mapFlowDataModelToRuleGoModel(
+    logicFlowData,
+    props.modelValue,
+  );
+  emit('update:modelValue', ruleGoData);
+}
 
 useResizeObserver(containerRef, (entries) => {
   const entry = entries[0];
@@ -797,12 +460,14 @@ onMounted(async () => {
 
 onBeforeUnmount(() => {
   window.removeEventListener('keydown', handleKeyDown);
+  handleEmitUpdate();// 组件卸载时触发更新事件
 });
 
 defineExpose({
   getData,
   clearFlowData,
   generateFlowData,
+  rerenderFlowData,
   openDrawer,
   closeDrawer,
 });
@@ -818,7 +483,7 @@ defineExpose({
       @update-node-properties="flowNodeUpdateNodePropertiesHandler"
       @selected="setSelectedNodeModel"
       @addComment="handleAddComment"
-    ></flow-view>
+    />
     <node-list
       ref="nodeListRef"
       :rootReact="containerReact"
@@ -826,7 +491,7 @@ defineExpose({
       :y="menuListY"
       :menu-list="menuList"
       @selected="nodeListSelectedHandler"
-    ></node-list>
+    />
     <node-form
       v-if="selectedNodeModel"
       ref="nodeFormRef"
@@ -837,9 +502,12 @@ defineExpose({
       :next-data="nextNodes"
       :anchor-can-connect-only-one-node="anchorCanConnectOnlyOneNode"
       :selected-node-id="selectedNodeModel.id"
-      :chain-id="props.flowData?.ruleChain?.id"
+      :chain-id="props.modelValue?.ruleChain?.id"
       @add="nodeFormAddHandler"
-    ></node-form>
-    <run-drawer ref="runDrawerRef" :flow-data="props.flowData"></run-drawer>
+    />
+    <run-drawer
+      ref="runDrawerRef"
+      :flow-data="props.modelValue"
+    />
   </div>
 </template>
